move pollhistory from dialogs to polls directory
