+++ res/css/views/polls/pollHistory/_PollHistoryDialog.pcss
