+++ res/css/views/polls/pollHistory/_PollHistoryList.pcss
