+++ src/components/views/polls/pollHistory/PollDetail.tsx
