+++ src/components/views/polls/pollHistory/PollDetailHeader.tsx
