+++ src/components/views/polls/pollHistory/PollHistoryList.tsx
