+++ src/components/views/polls/pollHistory/PollListItem.tsx
