+++ src/components/views/polls/pollHistory/PollListItemEnded.tsx
