+++ src/i18n/strings/en_EN.json
@@ -2319,6 +2319,22 @@
     "Verification cancelled": "Verification cancelled",
     "%(count)s votes|other": "%(count)s votes",
     "%(count)s votes|one": "%(count)s vote",
+    "View poll in timeline": "View poll in timeline",
+    "Active polls": "Active polls",
+    "Past polls": "Past polls",
+    "Loading polls": "Loading polls",
+    "Load more polls": "Load more polls",
+    "There are no active polls in this room": "There are no active polls in this room",
+    "There are no past polls in this room": "There are no past polls in this room",
+    "There are no active polls. Load more polls to view polls for previous months": "There are no active polls. Load more polls to view polls for previous months",
+    "There are no past polls. Load more polls to view polls for previous months": "There are no past polls. Load more polls to view polls for previous months",
+    "There are no active polls for the past %(count)s days. Load more polls to view polls for previous months|other": "There are no active polls for the past %(count)s days. Load more polls to view polls for previous months",
+    "There are no active polls for the past %(count)s days. Load more polls to view polls for previous months|one": "There are no active polls for the past day. Load more polls to view polls for previous months",
+    "There are no past polls for the past %(count)s days. Load more polls to view polls for previous months|other": "There are no past polls for the past %(count)s days. Load more polls to view polls for previous months",
+    "There are no past polls for the past %(count)s days. Load more polls to view polls for previous months|one": "There are no past polls for the past day. Load more polls to view polls for previous months",
+    "View poll": "View poll",
+    "Final result based on %(count)s votes|other": "Final result based on %(count)s votes",
+    "Final result based on %(count)s votes|one": "Final result based on %(count)s vote",
     "%(name)s started a video call": "%(name)s started a video call",
     "Video call ended": "Video call ended",
     "Sunday": "Sunday",
@@ -2412,8 +2428,6 @@
     "Vote not registered": "Vote not registered",
     "Sorry, your vote was not registered. Please try again.": "Sorry, your vote was not registered. Please try again.",
     "Due to decryption errors, some votes may not be counted": "Due to decryption errors, some votes may not be counted",
-    "Final result based on %(count)s votes|other": "Final result based on %(count)s votes",
-    "Final result based on %(count)s votes|one": "Final result based on %(count)s vote",
     "Results will be visible when the poll is ended": "Results will be visible when the poll is ended",
     "No votes cast": "No votes cast",
     "%(count)s votes cast. Vote to see the results|other": "%(count)s votes cast. Vote to see the results",
@@ -3129,20 +3143,6 @@
     "Not a valid Security Key": "Not a valid Security Key",
     "Access your secure message history and set up secure messaging by entering your Security Key.": "Access your secure message history and set up secure messaging by entering your Security Key.",
     "If you've forgotten your Security Key you can <button>set up new recovery options</button>": "If you've forgotten your Security Key you can <button>set up new recovery options</button>",
-    "View poll in timeline": "View poll in timeline",
-    "Active polls": "Active polls",
-    "Past polls": "Past polls",
-    "Loading polls": "Loading polls",
-    "Load more polls": "Load more polls",
-    "There are no active polls in this room": "There are no active polls in this room",
-    "There are no past polls in this room": "There are no past polls in this room",
-    "There are no active polls. Load more polls to view polls for previous months": "There are no active polls. Load more polls to view polls for previous months",
-    "There are no past polls. Load more polls to view polls for previous months": "There are no past polls. Load more polls to view polls for previous months",
-    "There are no active polls for the past %(count)s days. Load more polls to view polls for previous months|other": "There are no active polls for the past %(count)s days. Load more polls to view polls for previous months",
-    "There are no active polls for the past %(count)s days. Load more polls to view polls for previous months|one": "There are no active polls for the past day. Load more polls to view polls for previous months",
-    "There are no past polls for the past %(count)s days. Load more polls to view polls for previous months|other": "There are no past polls for the past %(count)s days. Load more polls to view polls for previous months",
-    "There are no past polls for the past %(count)s days. Load more polls to view polls for previous months|one": "There are no past polls for the past day. Load more polls to view polls for previous months",
-    "View poll": "View poll",
     "Send custom account data event": "Send custom account data event",
     "Send custom room account data event": "Send custom room account data event",
     "Event Type": "Event Type",
